+++ CHANGELOG.md
@@ -37,7 +37,7 @@ from contracts import contract
 ## 3.1.5
 
 - raise ValueError when moving component
-- ports have a cross_section
+- add cross_section property to ports
 - `gdsfactory/routing/fanout.py` passes cross_section settings from port into bend_s
 - fix manhattan text, avoid creating duplicated cells
 - fix pcm_optical